+++ src/cad2d_ir/codecs/__init__.py
@@ -3,6 +3,7 @@ from cad2d_ir.codecs.dxf import (
     dxf_to_ir,
     ir_to_dxf,
     read_dxf_file,
+    resolve_dxf_output_encoding,
     write_dxf_file,
 )
 
@@ -11,5 +12,6 @@ __all__ = [
     "dxf_to_ir",
     "ir_to_dxf",
     "read_dxf_file",
+    "resolve_dxf_output_encoding",
     "write_dxf_file",
 ]